OLD/api audio tryout/requirements.txt
google-genai>=1.0.0
yt-dlp>=2025.1.1
